requirements.txt
espn-api
python-dotenv
pandas 
requests 
tenacity
openai
mcp[cli]

# source venv/bin/activate